working CMake build on Linux
This required a few minor modifications. The LuaJIT find
CMake module uses different variables than the ones I was
using on Windows, so I renamed the ones I was using to
be the same as the ones in the module, which have no
mention of JIT allowing for easy replacing with the
vanilla Lua library (though I don't have intentions
of doing so).

A problem was that LuaJIT's FFI API uses the default
shared-library system loading procedure, so it wasn't
able to find the bindings shared object file,
for example, which was in the same directory as the
executable. It made no sense to install this library
since it's only used by this program, and it felt
hackish to add a wrapper script that sets the
LD_LIBRARY_PATH env-var. The FFI API does allow
loading a library from a specific path, but then it
no longer automatically formats the module name
(i.e. prepend lib and append .so) so this would have
introduced platform-dependent behavior which felt
unnecessary to me.

So I remembered about rpath linking and found out
about the $ORIGIN variable in the docs, which allows
one to specify an rpath relative to the executable.

Added the C++11 clang++ option.

Made the DLL copy post-build task be WIN32 only.

Added WIN32 conditional dllexport.

Also added a simple build script similar to the Windows
batch file.